+++ documentation/api/change_log.rst
@@ -8,6 +8,7 @@ API change log
 v3.0-19 | 2024-08-13
 """"""""""""""""""""
 
+- Allow passing a fixed price in the ``flex-context`` using the new fields ``consumption-price`` and ``production-price``, which are meant to replace the ``consumption-price-sensor`` and ``production-price-sensor`` fields, respectively.
 - Allow posting a single instantaneous belief as a list of one element to `/sensors/data` (POST).
 - Allow setting a SoC unit directly in some fields (formerly ``Float`` fields, and now ``Quantity`` fields), while still falling back on the contents of the ``soc-unit`` field, for backwards compatibility: